+++ R/FLStock.R
@@ -925,3 +925,11 @@ setMethod("sr", signature(object="FLStock"),
 
 		return(FLQuants(rec=rec, ssb=ssb))
 }) # }}}
+
+
+# catch.sel {{{
+setMethod("catch.sel", signature(object="FLStock"),
+  function(object) {
+    return(harvest(object) %/% apply(harvest(object), 2:6, max))
+  }
+) # }}}